Anomaly-detector/requirements.txt
a2wsgi==1.10.7
anyio
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
connexion==3.1.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools
httpx==0.27.2
idna==3.8
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kazoo==2.5.0
MarkupSafe==2.1.5
pykafka==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
swagger_ui_bundle==1.1.0
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop
watchfiles==0.24.0
websockets==13.0.1
Werkzeug==3.0.4
flask-cors
